TST: move orphaned business hour tests biz hour file (#27085)
